refactor: type annotate ReturnMapping
To facilitate this, we move the case where cls is None earlier, and
remove the handling of the error case, since the rpc will raise an
exception in the error case, which ReturnMapping doesn't catch, so the
error case logic won't be used anyway.

It looks like the sequence type case should be able to be removed too,
as it doesn't appear to be used. The None case could be removed entirely
by simply not adding the ReturnMapping decorator if cls is None (known at
codegen time).

Finally, we should be able to remove ReturnMapping entirely and just
have the decorated function directly do
return cls.from_json(reply['response'])