refactor: only make factories and classes that will be written
Move logic based on Facade being in the name to the point where we add
the definitions to be written to factories or definitions dicts. This
eliminates the need to skip writing definitions that have Facade in the
name or skip writing factories that don't have it -- we just only store
the ones we want to write.