fix: raise ConnectionClosed subtypes, correctly (+ update test raise)
Two arguments are required, rcvd and sent, both of which should be a
`websockets.frames.Closed` object or None. If neither are None, the
third argument should also be provided to say in which order sent and
rcvd occurred.

The existing style appears to have been based on an older API which
looks more similar to raising a `Closed` object directly (a code and
reason are provided).

The exception types are exposed on websockets directly, so I've also
removed references to websockets.exceptions.ConnectionClosed.